+++ lib/src/flutter_ble_peripheral.dart
@@ -10,7 +10,7 @@ import 'dart:typed_data';
 
 import 'package:flutter/services.dart';
 import 'package:flutter_ble_peripheral/src/models/advertise_data.dart';
-import 'package:flutter_ble_peripheral/src/models/advertise_error.dart';
+import 'package:flutter_ble_peripheral/src/models/advertise_exception.dart';
 import 'package:flutter_ble_peripheral/src/models/advertise_set_parameters.dart';
 import 'package:flutter_ble_peripheral/src/models/advertise_settings.dart';
 import 'package:flutter_ble_peripheral/src/models/enums/android_error.dart';
@@ -52,7 +52,10 @@ class FlutterBlePeripheral {
   //     'dev.steenbakker.flutter_ble_peripheral/ble_data_received');
 
   /// Start advertising. Takes [AdvertiseData] as an input.
-  Future<AdvertiseError?> start({
+  ///
+  /// Throws [Exception] if input is not valid.
+  /// Throws [AdvertiseException] if platform side error occurred.
+  Future<void> start({
     required AdvertiseData advertiseData,
     AdvertiseSettings? advertiseSettings,
     AdvertiseSetParameters? advertiseSetParameters,
@@ -126,17 +129,22 @@ class FlutterBlePeripheral {
 
     try {
       await _methodChannel.invokeMethod('start', parameters);
-    } on PlatformException catch(e) {
+    } on PlatformException catch (e) {
       final errorCode = int.tryParse(e.code) ?? 999;
       dynamic error;
       if (errorCode > 0 && errorCode <= 5) {
-        error = AndroidError.values.firstWhere((element) => element.code == errorCode);
+        error = AndroidError.values
+            .firstWhere((element) => element.code == errorCode);
       } else if (errorCode >= 10 && errorCode <= 16) {
-        error = PeripheralState.values.firstWhere((element) => element.code == errorCode);
+        error = PeripheralState.values
+            .firstWhere((element) => element.code == errorCode);
       }
-      return AdvertiseError(errorCode: errorCode, message: e.message, error: error);
+      throw AdvertiseException(
+        errorCode: errorCode,
+        message: e.message,
+        error: error,
+      );
     }
-    return null;
   }
 
   /// Stop advertising
@@ -164,7 +172,8 @@ class FlutterBlePeripheral {
 
   /// Request permission
   Future<bool> requestPermission() async {
-   return await _methodChannel.invokeMethod<bool>('requestPermission') ?? false;
+    return await _methodChannel.invokeMethod<bool>('requestPermission') ??
+        false;
   }
 
   /// Stop advertising