+++ lib/src/models/advertise_error.dart
@@ -1,7 +1,0 @@
-class AdvertiseError {
-  final int? errorCode;
-  final String? message;
-  final dynamic error;
-
-  AdvertiseError({this.errorCode, this.message, this.error});
-}