+++ lib/src/models/advertise_exception.dart
@@ -1,0 +1,7 @@
+class AdvertiseException implements Exception {
+  final int? errorCode;
+  final String? message;
+  final dynamic error;
+
+  AdvertiseException({this.errorCode, this.message, this.error});
+}